refactor: extract chat and file transport
